incorporated code from main

+++ pkg/sources/github/github.go
@@ -8,6 +8,7 @@ import (
 	"net/http"
 	"net/url"
 	"os"
+	"path/filepath"
 	"regexp"
 	"slices"
 	"sort"
@@ -741,7 +742,10 @@ func (s *Source) cloneAndScanRepo(ctx context.Context, repoURL string, repoInfo 
 	if err != nil {
 		return duration, err
 	}
-	defer os.RemoveAll(path)
+	// remove the path only if it was created as a temporary path, or if it is a clone path and --no-cleanup is not set.
+	if strings.HasPrefix(path, filepath.Join(os.TempDir(), "trufflehog")) || (!s.conn.NoCleanup && s.conn.GetClonePath() != "") {
+		defer os.RemoveAll(path)
+	}
 
 	// TODO: Can this be set once or does it need to be set on every iteration? Is |s.scanOptions| set every clone?
 	s.setScanOptions(s.conn.Base, s.conn.Head)
